Misc fixes and v3 support.
We are adding a new version of the symbolication protocol which will
support other operating systems. The only difference from v2 is that the
request has the computed breakpad id, not the pdb info that is used to
compute it on windows.

This changes adds support for it. It also includes

* A simple convert script which can be use for testing changes to firefox.
* Examples form a currently nightly (v2) and a local build (v3). I will update
  the v3 example with a nightly one once the firefox change is in one.
* Misc fixes/changes:
  * If a file has multiple hangs, we don't dump the rest of the payload multiple
    times.
  * If a file has no hangs, we still dump the rest of the payload once.