Fix parsing of lines with "special" leading charters.
The first character in each line can be value from 0 to 255. This is further
complicated by pig printing them as utf-8. To be able to parse this we cannot
simply read a line and split on tabs. We have to remember where in the file
we are and read the next expected token.

This pull request includes an extra example file with lines starting with \r,
\n, \t and a 3 byte utf-8 char.

With this change I was able to process a dump (Feb 15) without any errors.

Last but not least, this pull request removes the toplevel try..except that
just logs erros. We control this part of the format, and it is better to
know right away that something is wrong than to skip entries and produce
files with the wrong key.